Move more things to modulefinder.py
